Working on the ant.turn() method. So ants know what to do when it is their turn.
Stuck on logic for building weighted turn list. Need to change
isinstance(). Should be a check for pheremones, each pheremone
adds the point coords to the __moves__ list. Not sure how to
prioritise the food though, other than by giving it a large value
(100?), will probably be the same with hive.

If I do the food/hive after the pheremones I can check for the
most frequent coord and make sure to add more than that. May have
to add in a check to make sure it doesnt get stuck in some
stupidly large loop.

Tests will currently fail. :(